Move amount verification to main
